+++ catalyst-gateway/tests/api_tests/api_tests/__init__.py
@@ -1,5 +1,6 @@
 """Utilities for testing schema mismatch behavior."""
 
+from loguru import logger
 import http.client
 
 DB_URL = "postgres://catalyst-event-dev:CHANGE_ME@localhost/CatalystEventDev"
@@ -16,3 +17,21 @@ def call_api_url(method: str, endpoint: str):
     resp = client.getresponse()
     client.close()
     return resp
+
+
+def check_is_live():
+    resp = call_api_url("GET", "/api/health/live")
+    assert resp.status == 204
+    logger.info("cat-gateway service is LIVE.")
+
+
+def check_is_ready():
+    resp = call_api_url("GET", "/api/health/ready")
+    assert resp.status == 204
+    logger.info("cat-gateway service is READY.")
+
+
+def check_is_not_ready():
+    resp = call_api_url("GET", "/api/health/ready")
+    assert resp.status == 503
+    logger.info("cat-gateway service is NOT READY.")
