+++ catalyst-gateway/tests/api_tests/api_tests/test_utxo.py
@@ -1,0 +1,5 @@
+from api_tests import check_is_live
+
+
+def test_staked_ada_endpoint():
+    check_is_live()
